repaired views in planowanie treningu (needs to be added primary key)

+++ WebApplication/Controllers/PlanowanieTreningowController.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +12,7 @@ using WebApplication.Models;
 
 namespace WebApplication.Controllers
 {
+    [Authorize]
     public class PlanowanieTreningowController : Controller
     {
         private readonly MyContext _context;
@@ -22,7 +25,8 @@ namespace WebApplication.Controllers
         // GET: PlanowanieTreningow
         public async Task<IActionResult> Index()
         {
-            var myContext = _context.planowaneTreningi.Include(p => p.trening).Include(p => p.uzytkownik);
+            int userId = int.Parse(User.Identity.GetUserId());
+            var myContext = _context.planowaneTreningi.Include(p => p.trening).Where(x => x.id_uzytkownika == userId );
             return View(await myContext.ToListAsync());
         }
 
@@ -43,6 +47,9 @@ namespace WebApplication.Controllers
                 return NotFound();
             }
 
+            if (planowanieTreningow.id_uzytkownika != int.Parse(User.Identity.GetUserId()))
+                return RedirectToAction("Index");
+
             return View(planowanieTreningow);
         }
 
@@ -51,7 +58,6 @@ namespace WebApplication.Controllers
         public IActionResult Create(string id_string)
         {
             ViewData["id_treningu"] = new SelectList(_context.treningi, "id_treningu", "nazwa");
-            ViewData["id_uzytkownika"] = new SelectList(_context.uzytkownicy, "Id", "Id");
             if (!String.IsNullOrEmpty(id_string))
             {
                 int id = int.Parse(id_string);
@@ -74,16 +80,16 @@ namespace WebApplication.Controllers
         {
             if (ModelState.IsValid)
             {
+                planowanieTreningow.id_uzytkownika = int.Parse(User.Identity.GetUserId());
                 _context.Add(planowanieTreningow);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
             ViewData["id_treningu"] = new SelectList(_context.treningi, "id_treningu", "nazwa", planowanieTreningow.id_treningu);
-            ViewData["id_uzytkownika"] = new SelectList(_context.uzytkownicy, "Id", "Id", planowanieTreningow.id_uzytkownika);
             return View(planowanieTreningow);
         }
 
-        // GET: PlanowanieTreningow/Edit/5
+        /*/ GET: PlanowanieTreningow/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -136,7 +142,7 @@ namespace WebApplication.Controllers
             ViewData["id_treningu"] = new SelectList(_context.treningi, "id_treningu", "nazwa", planowanieTreningow.id_treningu);
             ViewData["id_uzytkownika"] = new SelectList(_context.uzytkownicy, "Id", "Id", planowanieTreningow.id_uzytkownika);
             return View(planowanieTreningow);
-        }
+        }*/
 
         // GET: PlanowanieTreningow/Delete/5
         public async Task<IActionResult> Delete(int? id)
@@ -155,6 +161,9 @@ namespace WebApplication.Controllers
                 return NotFound();
             }
 
+            if (planowanieTreningow.id_uzytkownika != int.Parse(User.Identity.GetUserId()))
+                return RedirectToAction("Index");
+
             return View(planowanieTreningow);
         }
 
@@ -164,6 +173,10 @@ namespace WebApplication.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var planowanieTreningow = await _context.planowaneTreningi.FindAsync(id);
+
+            if (planowanieTreningow.id_uzytkownika != int.Parse(User.Identity.GetUserId()))
+                return RedirectToAction("Index");
+
             _context.planowaneTreningi.Remove(planowanieTreningow);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
